miners module refactor
