--- neurons/miners/wiki_agent/requirements.txt
@@ -1,5 +1,0 @@
-# If so, we need to make it clear on the README. Otherwise, we should have a completely separated requirements for the miner
-openai==0.28
-langchain==0.1.0
-python-dotenv
-wikipedia
