+++ prompting/miners/__init__.py
@@ -1,0 +1,10 @@
+# Test miners
+from .echo import EchoMiner
+from .mock import MockMiner
+from .phrase import PhraseMiner
+
+# Real miners
+from .hf_miner import HuggingFaceMiner
+from .openai_miner import OpenAIMiner
+from .wiki_agent_miner import WikipediaAgentMiner
+
