+++ requirements.txt
@@ -16,3 +16,6 @@ wandb==0.15.10
 tenacity
 antlr4-python3-runtime==4.11
 wikipedia
+openai==1.9.0
+langchain==0.1.0
+python-dotenv
